requirements.txt
streamlit
google-generativeai
plotly
python-dotenv
reportlab
pandas